Connectivity refactoring
