+++ src/ConnectivityWizard/Connectivity/Connectivity.ts
@@ -1,0 +1,3 @@
+export interface Connectivity {
+    test(): Promise<boolean>;
+};
