+++ src/ConnectivityWizard/Entities/ConnectivityStatus.ts
