+++ src/ConnectivityWizard/Entities/ConnectivityTest.ts
