+++ src/ConnectivityWizard/Entities/ScanId.ts
