+++ src/ConnectivityWizard/Entities/ScannedUrl.ts
